pyproject.toml
[project]
name = "techread"
version = "0.1.0"
description = "Local CLI tool to fetch, rank, and summarize technical blogs."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name="techread contributors"}]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "httpx>=0.27.0",
  "feedparser>=6.0.11",
  "trafilatura>=1.8.1",
  "python-dateutil>=2.9.0.post0",
]

[project.scripts]
techread = "techread.cli:app"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
